start migrating languages call to opnestad-global
