+++ packages/cms/lib/modules/openstad-global/lib/arrangeFields.js
@@ -43,8 +43,15 @@ module.exports = [
     {
         name: 'mainMenu',
         label: 'Hoofdmenu',
-        fields: ['arrangeMainMenu', 'displayLoginTopLink', 'mainMenuItems', 'ctaButtonText', 'ctaButtonUrl', 'topLinks', 'displayMyAcount', "myAccountButtonText", 'shouldAutoTranslate']
+        fields: ['arrangeMainMenu', 'displayLoginTopLink', 'mainMenuItems', 'ctaButtonText', 'ctaButtonUrl', 'topLinks', 'displayMyAcount', "myAccountButtonText", 'shouldAutoTranslate', ]
     },
+    {
+        name: 'translations',
+        label: 'Translating',
+        fields: ['deeplTranslationKey' ]
+    },
+
+    
     {
         name: 'userRights',
         label: 'Roles & Rights',
